Improve server logging The clicks.txt file moved into the directory `dynamic`. This is so that this folder can be shared between the host and the docker container. In the future (the next couple commits) the logger will be configured to also write the log files into that folder.
All `println!` and similar got replaced by macros from the `log` crate.

The logging targets got changed from `all` for everything to `web` for
web specific things and `server` for internal server information.

The production logger format function now also prints this target. This
way it is easy to write a script to only filter the Web specific logs.

The debug logger format function now checks the target and prevents a
panic when it tries to split the string.