Logger now save logs to file The server now creates log-files and saves them into `dynamic/logs`. It also still writes the logs to stderr.
The log-files can grow to a max of 50MB and there is a hard limit of 100
log files. So there may never be more than 5GB of log-files on the
server.
Those configurations can (and probably will) be changed in the future.

Also there is now a simple shell script called `delete_logs.sh` to
delete all logs.